+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `IChannelMethods` interface to deprecate `IPreviewChannelMethods` and `IOnlineChannelMethods` as since version 2.0.0 they were the same
 
 ## [2.2.3] - 2020-05-27
 ### Added